resource based authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 
@@ -9,7 +11,8 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishes
 {
     public class DeleteDishesForRestaurantCommandHandler(IRestaurantsRepository restaurantsRepository,
         IDishesRepository dishesRepository,
-        ILogger<DeleteDishesForRestaurantCommandHandler> logger)
+        ILogger<DeleteDishesForRestaurantCommandHandler> logger,
+        IRestaurantAuthorizationService restaurantAuthorizationService)
         : IRequestHandler<DeleteDishesForRestaurantCommand>
     {
         public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
@@ -20,6 +23,13 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishes
             {
                 throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
             }
+
+            //check authorization
+            if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+            {
+                throw new ForbidException();
+            }
+
             await dishesRepository.Delete(restaurant.Dishes);
         }
     }
