+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -2,7 +2,9 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Application.Restaurants.Commands.CreateRestaurant;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 using System;
 using System.Collections.Generic;
@@ -14,7 +16,8 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
 {
     public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger,
         IMapper mapper,
-        IRestaurantsRepository restaurantsRepository
+        IRestaurantsRepository restaurantsRepository,
+        IRestaurantAuthorizationService restaurantAuthorizationService
         ) : IRequestHandler<DeleteRestaurantCommand>
     {
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
@@ -23,6 +26,13 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
             var restaurant = await restaurantsRepository.GetById(request.Id);
             if (restaurant is null)
                 throw new NotFoundException(nameof(restaurant), request.Id.ToString());
+
+            //check authorization
+            if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+            {
+                throw new ForbidException();
+            }
+
             await restaurantsRepository.Delete(restaurant);
 
         }
